探索-add-还原

+++ Assets/Script/Game/Module/DataCenter/Components/ExploreData.cs
@@ -73,7 +73,7 @@ namespace SGame
 				}
 				c_cache_attr = ls;
 				c_strong_power = GlobalDesginConfig.GetFloat("battle_equip_fortify", 2f);
-				c_exploretool_time_cost = GlobalDesginConfig.GetFloat("battle_explore_gem", 100); 
+				c_exploretool_time_cost = GlobalDesginConfig.GetInt("battle_explore_gem", 100); 
 
 				data.Refresh();
 			}
